eliminate validators
